Agent model settings

+++ src/ycl_aaic_agent_db.clas.abap
@@ -19,7 +19,7 @@ ENDCLASS.
 
 
 
-CLASS YCL_AAIC_AGENT_DB IMPLEMENTATION.
+CLASS ycl_aaic_agent_db IMPLEMENTATION.
 
 
   METHOD yif_aaic_agent_db~create.
@@ -38,18 +38,8 @@ CLASS YCL_AAIC_AGENT_DB IMPLEMENTATION.
     ENDIF.
 
     DATA(ls_agent) = i_s_agent.
-    DATA(lt_agent_tools) = i_t_agent_tools.
-    DATA(lt_agent_models) = i_t_agent_models.
 
     ls_agent-id = xco_cp=>uuid( )->value.
-
-    LOOP AT lt_agent_tools ASSIGNING FIELD-SYMBOL(<ls_agent_tool>).
-      <ls_agent_tool>-id = ls_agent-id.
-    ENDLOOP.
-
-    LOOP AT lt_agent_models ASSIGNING FIELD-SYMBOL(<ls_agent_model>).
-      <ls_agent_model>-id = ls_agent-id.
-    ENDLOOP.
 
     INSERT yaaic_agent FROM @ls_agent.
 
@@ -60,17 +50,45 @@ CLASS YCL_AAIC_AGENT_DB IMPLEMENTATION.
 
     e_id = ls_agent-id.
 
-    INSERT yaaic_agent_tool FROM TABLE @lt_agent_tools.
+    IF i_t_agent_tools IS SUPPLIED.
 
-    IF sy-subrc <> 0.
-      e_error = |Error while saving tools for Agent { i_s_agent-name }|.
-      RETURN.
+      DATA(lt_agent_tools) = i_t_agent_tools.
+
+      LOOP AT lt_agent_tools ASSIGNING FIELD-SYMBOL(<ls_agent_tool>).
+        <ls_agent_tool>-id = ls_agent-id.
+      ENDLOOP.
+
+      IF lt_agent_tools IS NOT INITIAL.
+
+        INSERT yaaic_agent_tool FROM TABLE @lt_agent_tools.
+
+        IF sy-subrc <> 0.
+          e_error = |Error while saving tools for Agent { i_s_agent-name }|.
+          RETURN.
+        ENDIF.
+
+      ENDIF.
+
     ENDIF.
 
-    INSERT yaaic_agent_mdl FROM TABLE @lt_agent_models.
+    IF i_t_agent_models IS SUPPLIED.
 
-    IF sy-subrc <> 0.
-      e_error = |Error while saving models for Agent { i_s_agent-name }|.
+      DATA(lt_agent_models) = i_t_agent_models.
+
+      LOOP AT lt_agent_models ASSIGNING FIELD-SYMBOL(<ls_agent_model>).
+        <ls_agent_model>-id = ls_agent-id.
+      ENDLOOP.
+
+      IF lt_agent_models IS NOT INITIAL.
+
+        INSERT yaaic_agent_mdl FROM TABLE @lt_agent_models.
+
+        IF sy-subrc <> 0.
+          e_error = |Error while saving models for Agent { i_s_agent-name }|.
+        ENDIF.
+
+      ENDIF.
+
     ENDIF.
 
   ENDMETHOD.
@@ -133,30 +151,40 @@ CLASS YCL_AAIC_AGENT_DB IMPLEMENTATION.
     e_updated = abap_true.
 
     DELETE FROM yaaic_agent_tool WHERE id = @i_s_agent-id.
-    DELETE FROM yaaic_agent_mdl WHERE id = @i_s_agent-id.
 
-    DATA(lt_agent_tools) = i_t_agent_tools.
-    DATA(lt_agent_models) = i_t_agent_models.
+    IF i_t_agent_tools IS SUPPLIED.
 
-    LOOP AT lt_agent_tools ASSIGNING FIELD-SYMBOL(<ls_agent_tool>).
-      <ls_agent_tool>-id = i_s_agent-id.
-    ENDLOOP.
+      DATA(lt_agent_tools) = i_t_agent_tools.
 
-    LOOP AT lt_agent_models ASSIGNING FIELD-SYMBOL(<ls_agent_model>).
-      <ls_agent_model>-id = i_s_agent-id.
-    ENDLOOP.
+      LOOP AT lt_agent_tools ASSIGNING FIELD-SYMBOL(<ls_agent_tool>).
+        <ls_agent_tool>-id = i_s_agent-id.
+      ENDLOOP.
 
-    INSERT yaaic_agent_tool FROM TABLE @lt_agent_tools.
+      INSERT yaaic_agent_tool FROM TABLE @lt_agent_tools.
 
-    IF sy-subrc <> 0.
-      e_error = |Error while saving tools for Agent { i_s_agent-name }|.
-      RETURN.
+      IF sy-subrc <> 0.
+        e_error = |Error while saving tools for Agent { i_s_agent-name }|.
+        RETURN.
+      ENDIF.
+
     ENDIF.
 
-    INSERT yaaic_agent_mdl FROM TABLE @lt_agent_models.
+    IF i_t_agent_models IS SUPPLIED.
 
-    IF sy-subrc <> 0.
-      e_error = |Error while saving models for Agent { i_s_agent-name }|.
+      DELETE FROM yaaic_agent_mdl WHERE id = @i_s_agent-id.
+
+      DATA(lt_agent_models) = i_t_agent_models.
+
+      LOOP AT lt_agent_models ASSIGNING FIELD-SYMBOL(<ls_agent_model>).
+        <ls_agent_model>-id = i_s_agent-id.
+      ENDLOOP.
+
+      INSERT yaaic_agent_mdl FROM TABLE @lt_agent_models.
+
+      IF sy-subrc <> 0.
+        e_error = |Error while saving models for Agent { i_s_agent-name }|.
+      ENDIF.
+
     ENDIF.
 
   ENDMETHOD.
